\name{akc-package}
\alias{akc-package}
\alias{akc}
\docType{package}
\title{
\packageTitle{akc}
}
\description{
\packageDescription{akc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{akc}
\packageIndices{akc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{akc}

Maintainer: \packageMaintainer{akc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
