% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_export.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{group_components}
\alias{group_edge_betweenness}
\alias{group_fast_greedy}
\alias{group_infomap}
\alias{group_label_prop}
\alias{group_leading_eigen}
\alias{group_louvain}
\alias{group_optimal}
\alias{group_spinglass}
\alias{group_walktrap}
\alias{group_biconnected_component}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tidygraph}{\code{\link[tidygraph:group_graph]{group_biconnected_component}}, \code{\link[tidygraph:group_graph]{group_components}}, \code{\link[tidygraph:group_graph]{group_edge_betweenness}}, \code{\link[tidygraph:group_graph]{group_fast_greedy}}, \code{\link[tidygraph:group_graph]{group_infomap}}, \code{\link[tidygraph:group_graph]{group_label_prop}}, \code{\link[tidygraph:group_graph]{group_leading_eigen}}, \code{\link[tidygraph:group_graph]{group_louvain}}, \code{\link[tidygraph:group_graph]{group_optimal}}, \code{\link[tidygraph:group_graph]{group_spinglass}}, \code{\link[tidygraph:group_graph]{group_walktrap}}}
}}

