% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{English stop words collected in tidytext package}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1149 rows and 2 columns.}
\usage{
stop_words
}
\description{
See \code{\link[tidytext]{stop_words}} from \pkg{tidytext} package.
}
\keyword{datasets}
