% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akc-package.R
\docType{package}
\name{akc-package}
\alias{akc}
\alias{akc-package}
\title{akc: Automatic Knowledge Classification}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A tidy framework for automatic knowledge classification and visualization based on keyword co-occurrence network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hope-data-science/akc}
}

}
\author{
\strong{Maintainer}: Tian-Yuan Huang \email{huang.tian-yuan@qq.com} (\href{https://orcid.org/0000-0002-3591-4203}{ORCID})

}
\keyword{internal}
