% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akc-package.R
\docType{package}
\name{akc-package}
\alias{akc}
\alias{akc-package}
\title{akc: Automatic Knowledge Classification}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tidy framework for automatic knowledge classification and visualization. Currently, the core functionality of the framework is mainly supported by modularity-based clustering (community detection) in keyword co-occurrence network, and focuses on co-word analysis of bibliometric research. However, the designed functions in 'akc' are general, and could be extended to solve other tasks in text mining as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hope-data-science/akc}
}

}
\author{
\strong{Maintainer}: Tian-Yuan Huang \email{huang.tian-yuan@qq.com} (\href{https://orcid.org/0000-0002-4151-3764}{ORCID})

}
\keyword{internal}
